(PC-12694) rename SearhcButton into Button
