+++ adage-front/src/app/ui-kit/Button/Button.scss
@@ -1,7 +1,7 @@
 @import 'styles';
 @import 'sass-rem';
 
-.search-button {
+.button {
   border-style: none;
   border-radius: rem(24px);
   padding: rem(10px) rem(24px);
