--- adage-front/src/app/ui-kit/SearchButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchButton'
