+++ adage-front/src/app/ui-kit/index.ts
@@ -1,4 +1,5 @@
 export { default as Checkbox } from './Checkbox'
-export { default as MultiSelectAutoComplete } from './MultiSelectAutoComplete'
+export { default as MultiSelectAutocomplete } from './MultiSelectAutoComplete'
 export { default as Tag } from './Tag'
 export { default as Pellet } from './Pellet'
+export { default as Button } from './Button'
